requirements.txt
fastapi
uvicorn
pydantic
faststream[rabbit]
motor
redis
python-dotenv
email-validator